Upload of answers
In this Repository i have upload 3 files. In which two file of tableau and one is just a normal txt file. all Answer contain in the one tableau file which is "tableau Exam Answer (Example of tableau)" except Q4. For the Question 4 i have created a separate Tableau file. and All the explanation of the all Question is present in the "Answers.txt" file.
Thank You!!